Fix confusing UI
Some spells have multiple damage components. For instance, Ahri's charm. But visually, this may be confusing to the user to see a spell appear twice in the analysis. To fix this we simply create a general damage analysis of a spell that is the sum of each damage part and keep track of each part for future usage for a deeper analysis.